Update README; Make window structs inherit fyne.Window and add fyne.Window to window interfaces; Move windows into windows package, including constants
